requirements.txt
Django==5.0.6
djangorestframework==3.15.0
mysqlclient==2.2.4
django-mysql==4.13.0
django-cors-headers==4.3.1
pika==1.3.2
djangorestframework_simplejwt==5.3.1
python-dotenv
pymongo==3.11.0
django==3.2.12
requests==2.28.2